% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_IRM.R
\name{batch_IRM}
\alias{batch_IRM}
\title{Run the core_IRM multiple times}
\usage{
batch_IRM(batch_list, rf_list, irf_list, match_agents)
}
\arguments{
\item{batch_list}{a list of configuration matrices to use as inputs into the core_IRM}

\item{rf_list}{the list of pathway data frames.}

\item{irf_list}{the list of irf data frames.}

\item{match_agents}{a logical value indicating whether the rf agent must match the irf, default set to FALSE}
}
\description{
\code{batch_IRM} run the core_IRM multiple times and return a list of three tibbles,
}
