% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions_funcs.R
\name{check_agent}
\alias{check_agent}
\title{Ensure data for a single RF agent is used on convolution}
\usage{
check_agent(irf, rf, match_agents = FALSE)
}
\arguments{
\item{irf}{the impulse response function tibble, must contain a year and value column}

\item{rf}{the radiative forcing time series tibble, must contain a year and value column}

\item{match_agents}{a logical value indicating whether to throw an error if the rf agent does not match the irf agent, default value is set to FALSE}
}
\value{
an error message if multiple agent conditions are not met in the rf and irf tibbles
}
\description{
\code{check_agent} make sure that the rf and irf tibbles contain data for a single RF agent, if match_agents
is true then make sure that both the RF agent in both rf and irf tibbles are the same.
}
\keyword{internal}
