% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions_funcs.R
\name{check_columns}
\alias{check_columns}
\title{Ensures that an object contains the required columns.}
\usage{
check_columns(object, required_columns, message_details = NULL)
}
\arguments{
\item{object}{the data frame or tibble to check}

\item{required_columns}{names of the columns that must be included in the object}

\item{message_details}{default set to NULL but if the user sets equal to a string, the contents of the string will be included in the error message}
}
\value{
an error message if a required column is missing
}
\description{
\code{check_columns} if the object is missing a required column then this function throws an error. If the
object contains all or more of the required columns the object passes this condition check.
}
\keyword{internal}
