% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions_funcs.R
\name{check_config_matrix}
\alias{check_config_matrix}
\title{Check the configuration matrix used in core_IRM}
\usage{
check_config_matrix(m)
}
\arguments{
\item{m}{configuration matrix to check}
}
\value{
an error message if the one of the matrix conditions are not met by the matrix
}
\description{
\code{check_config_matrix} make sure that the configuration matrix meets the following conditions:
it is a matrix; that is contains binary entires; that is contains at least one entry equal to one
otherwise the contributing_temp wont' run; and that the matrix has column and row names.
at least one entry that is equal to one otherwise contributing_temp won't run, and a
}
\keyword{internal}
