% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_conditions_funcs.R
\name{check_time}
\alias{check_time}
\title{Check the irf and rf inputs before performing convolution}
\usage{
check_time(irf, rf)
}
\arguments{
\item{irf}{the impulse response function tibble, must contain a year and value column}

\item{rf}{the radiative forcing time series tibble, must contain a year and value column}
}
\value{
an error message if the two data frames do not meet the criteria.
}
\description{
\code{check_time} checks to see if the IRF and the RF time series can be be convolved in
the \code{contributing_temp} function. The irf and rf data frame must contain values that correspond to
temporal data that consists of only one time step, that is equal in both data frames and spans an equal number of years.
If thes conditions are not met then this function will throw an error and prevent the convolution.
}
\keyword{internal}
