% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_IRM.R
\name{contributing_temp}
\alias{contributing_temp}
\title{The temperature contribtuion of a single RF agent}
\usage{
contributing_temp(rf, irf, match_agents)
}
\arguments{
\item{rf}{a tibble of a RF time series for a single RF agent, must contain a year, value, name, and agent column}

\item{irf}{a tibble of the IRF to convolve wihth the RF time series, must contain a year, value, name, and agent column}

\item{match_agents}{a logial value indicating whether the rf agent must match the irf, default set to FALSE}
}
\value{
a tibble containting a year, value, varibale, RF_name, IRF_name, and agent columns
}
\description{
\code{contributing_temp} convolves a RF timeseries with an IRF for a single radiative forcing agent, this is
equivalent to the temperature contirbution of a single radiative forcing agent. This function is used in \code{core_IRM}
}
\keyword{internal}
