% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_IRM.R
\name{core_IRM}
\alias{core_IRM}
\title{IRM}
\usage{
core_IRM(config_matrix, rf_list, irf_list, match_agents)
}
\arguments{
\item{config_matrix}{a matrix containing boolean indicators that refers to a specific RF time series and IRF to use as inputs in the contributing_temp function}

\item{rf_list}{a list of the RF time series}

\item{irf_list}{a list of the IRF}

\item{match_agents}{a logial value indicating whether the rf agent must match the irf, default set to FALSE}
}
\value{
a list of three tibbles, one of the total temp, one of the total RF, and one containting the contibuting temp from each RF agent
}
\description{
\code{core_IRM} is the IRM, calculates total temperature
}
