% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hector_comparison}
\alias{Hector_comparison}
\title{Hector_comparison}
\format{A list of a data frame containing
\describe{
#' \item{scenario}{The name of the Hector scenario run}
\item{year}{Interger value of the year}
\item{variable}{A description of the value Tgav (global mean temperature anomaly)}
\item{units}{A string of units for value (degC or W/m2)}
}}
\usage{
Hector_comparison
}
\description{
This is a list of the Hector temperature anomaly that is used as comparison
data in the HIRM validation experiments.
}
\concept{Comparison data}
\keyword{datasets}
