% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hector_ConfigMatrix}
\alias{Hector_ConfigMatrix}
\title{Hector_ConfigMatrix}
\format{A matrix with named rows and columns.
\describe{
\item{rows}{The rows are named after the RF drivers.}
\item{columns}{The columns are named after the IRFs.}
}}
\usage{
Hector_ConfigMatrix
}
\description{
This is an empty input configuration matrix for Hector.
}
\section{Notes}{
 The core HRIM configuration matrix is set up with RF drivers as rows and IRF as columns.
Right now the matix is empty, and only contains the value 0. When values in the matrix are set to 1 then
the RF driver and the IRF paring corresponding to the value will be used in HIRM.
}

\concept{Configuration matrices}
\keyword{datasets}
