% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hector_IRF}
\alias{Hector_IRF}
\title{Hector_IRF}
\format{An object of class \code{list} of length 1.}
\usage{
Hector_IRF
}
\description{
This is a list of the Hector impulse response functions.
Each element in the list should match the impulse response function.
}
\seealso{
Other IRF inputs: \code{\link{Sand_BC_IRF}}
}
\concept{IRF inputs}
\keyword{datasets}
