% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hector_RF}
\alias{Hector_RF}
\title{Hector_RF}
\format{A list of a data frame containing the RF driver time series.
\describe{
\item{scenario}{The name of the Hector core the output came from.}
\item{year}{Absolute year.}
\item{variable}{The name of the value, a type of RF.}
\item{value}{The RF value.}
\item{units}{The units for the RF value.}
\item{name}{The name of the RF driver.}
\item{agent}{The name of the RF agent.}
}}
\usage{
Hector_RF
}
\description{
This is a list of Hector RF drivers.
}
\concept{RF inputs}
\keyword{datasets}
