% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sand_BC_IRF}
\alias{Sand_BC_IRF}
\title{Sand_BC_IRF}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3001 rows and 5 columns.}
\usage{
Sand_BC_IRF
}
\description{
This is a dataframe of the BC IRF derived from the temperature response to a BC step test from Sand et al.
}
\seealso{
Other IRF inputs: \code{\link{Hector_IRF}}
}
\concept{IRF inputs}
\keyword{datasets}
